refactor(tests/client): move `FHIRcast` tests to new file
