+++ packages/core/src/client-test-utils.ts
@@ -1,0 +1,23 @@
+import { OperationOutcome } from '@medplum/fhirtypes';
+import { FetchLike } from './client';
+import { ContentType } from './contenttype';
+import { getStatus, isOperationOutcome } from './outcomes';
+
+export function mockFetch(
+  status: number,
+  body: OperationOutcome | Record<string, unknown> | ((url: string, options?: any) => any),
+  contentType = ContentType.FHIR_JSON
+): FetchLike & jest.Mock {
+  const bodyFn = typeof body === 'function' ? body : () => body;
+  return jest.fn((url: string, options?: any) => {
+    const response = bodyFn(url, options);
+    const responseStatus = isOperationOutcome(response) ? getStatus(response) : status;
+    return Promise.resolve({
+      ok: responseStatus < 400,
+      status: responseStatus,
+      headers: { get: () => contentType },
+      blob: () => Promise.resolve(response),
+      json: () => Promise.resolve(response),
+    });
+  });
+}
